==> code/2.1_select_baseline_characteristics.Rmd <==
---
title: "1.0_subset_baseline characteristics"
author: "Mark_Cheng"
date: "2024-01-18"
output: html_document
---


```{r setup, include=FALSE}
# load libraries
if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr, tidyr, stringr)

data_directory = "../data/"
data_directory_files = paste0(data_directory,list.files(data_directory))

DE_csv = data_directory_files[str_detect(data_directory_files,
                                         "_DE_knownDx_peerReviewed.csv")] # generated by 0_cleaning.Rmd

cleaned_df <- read.csv(DE_csv,header = TRUE, check.names = FALSE)
df_location_date <- read.csv("../output/Timeline_and_Lineage_Imputation/Studies_split_Timeperiods_ordered_by_time_periods.csv",check.names = FALSE) # generated from 1_Dominant_Lineage_assignment.Rmd
```

```{r} 
# order studies
df_location_date <- read.csv("../output/Timeline_and_Lineage_Imputation/Studies_split_CountriesAndTimeperiods_ordered_by_time_periods.csv",check.names = FALSE)
df_location_date

df_location_date %>% select(`Covidence #`) %>% unique()
```


```{r}
study_summary <- cleaned_df %>% select (`Covidence #`,`Study ID`,
                       `Peer-reviewed or Preprint`,Journal,`Database/Study ID`,
                       `Date of Publication (Month and Year)`, `Retrospective/Prospective`,
                       `Study Design`,`Study Country/Countries`,
                       `Study Start Date`, `Study End Date`,
                       `Total study size (number of patients)`,
                       `Primary Outcomes of Study`,`Secondary outcomes of study`,
                       `Cancer Total number`,`Control (non-cancer) Total number`,
                       `Explicitly included vaccinated patients`,
                       `outcomes described`,`Definition of Cancer patient`,
                       `Definition of severity`,`Definition of severe disease`,`Definition of ICU admission/Critical`,`Definition of mortality`,
                       `Univariate or multivariate analysis?`,`Multivariate analysis is adjusted by`,
                       `Comparison group`,`Comparator group matching strategy`,
                       `Which COVID-19 variant does the study identify?`,
                       c(`Covidence #`,`Study ID`,
                        colnames(cleaned_df)[grepl(pattern ="Age",colnames(cleaned_df))],
                        colnames(cleaned_df)[grepl(pattern ="Sex",colnames(cleaned_df))],
                        colnames(cleaned_df)[grepl(pattern ="Total number$",colnames(cleaned_df))]
                        )
                       )
study_summary
# reorder by order in figure 1
study_summary <- study_summary[order(match(study_summary$`Covidence #`,df_location_date$`Covidence #`)),]
study_summary
write.csv(study_summary,file = "../output/study_summary.csv",row.names = FALSE)
```
